flowertune-tinybert-lora/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowertune-tinybert-lora"
version = "1.0.0"
description = "FlowerTune TinyBERT LoRA: Federated TinyBERT LoRA Fine-tuning with Flower"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.26.1",
    "flwr-datasets>=0.6.0",
    "torch==2.8.0",
    "bitsandbytes==0.49.1",
    "scipy==1.13.0",
    "peft==0.6.2",
    "fschat[model_worker,webui]==0.2.35",
    "transformers==4.53.0",
    "sentencepiece==0.2.1",
    "omegaconf==2.3.0",
    "hf_transfer==0.1.8",
    "evaluate>=0.4.6",
    "scikit-learn>=1.8.0",
    "wandb>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "flowertune_tinybert_lora.server_app:app"
clientapp = "flowertune_tinybert_lora.client_app:app"

[tool.flwr.app.config]
dataset.name = "nyu-mll/glue"
dataset.subset = "qnli"
model.name = "huawei-noah/TinyBERT_General_4L_312D"
model.quantization = 8
model.gradient-checkpointing = false
model.lora.peft-lora-r = 8
model.lora.peft-lora-alpha = 16
train.save-every-round = 5
train.learning-rate-max = 5e-4
train.learning-rate-min = 1e-5
train.training-arguments.output-dir = ""
train.training-arguments.learning-rate = ""
train.training-arguments.per-device-train-batch-size = 32
train.training-arguments.per-device-eval-batch-size = 32
train.training-arguments.logging-steps = 10
train.training-arguments.num-train-epochs = 3
train.training-arguments.max-steps = 10
train.training-arguments.save-steps = 1000
train.training-arguments.save-total-limit = 10
train.training-arguments.gradient-checkpointing = false
train.training-arguments.lr-scheduler-type = "constant"
strategy.fraction-train = 0.2
strategy.fraction-evaluate = 0.0
num-server-rounds = 15
num-supernodes = 50
wandb.project = "flowertune-tinybert-lora"
wandb.entity = "ilham-abdillah-alhamdi-universitas-indonesia"

[superlink]
default = "local"

[superlink.local]
options.num-supernodes = 50
options.backend.client-resources.num-cpus = 1
options.backend.client-resources.num-gpus = 0.0